+++ packages/create-app/template/waku/src/pages/(home)/_layout.tsx
@@ -5,9 +5,3 @@ import { HomeLayout } from 'fumadocs-ui/layouts/home';
 export default function Layout({ children }: { children: ReactNode }) {
   return <HomeLayout {...baseOptions()}>{children}</HomeLayout>;
 }
-
-export const getConfig = async () => {
-  return {
-    render: 'static',
-  };
-};